backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
PyMySQL==1.1.1
Werkzeug==3.1.3
python-dotenv==1.1.0
marshmallow==3.25.1
